This device is for human safety
The statement of our problem is we’re creating a device which will spray the room while a person will enter there besides that it can spread the UV light for the 99.99% cleanliness. As we know the UV ray can demolish the coronavirus. It can also destroy the virus if it is contaminated with the air.
So, this device will spray UV ray in absence of the human.